borealis/requirements.txt
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
h11==0.16.0
httptools==0.7.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
prompt_toolkit==3.0.52
pyconfigstore3==1.0.1
pyfiglet==1.0.4
Pygments==2.20.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
questionary==2.1.1
regex==2026.4.4
requests==2.33.1
rich==14.3.4
setuptools==82.0.1
six==1.17.0
termcolor==3.3.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
